+++ rullst-capital/src/providers/mod.rs
@@ -21,6 +21,7 @@ pub mod razorpay;
 mod razorpay_webhook;
 pub mod stripe;
 mod stripe_charge;
+mod stripe_checkout;
 mod stripe_subscription;
 mod stripe_usage;
 pub mod wise;